% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCauses.R
\name{checkCauses}
\alias{checkCauses}
\title{Check availability of a cause in competing risk settings}
\usage{
checkCauses(cause, object)
}
\arguments{
\item{cause}{cause of interest}

\item{object}{object either obtained with \code{Hist} or \code{prodlim}}
}
\description{
For competing risk settings, check if the requested cause is known to the object
}
