% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{head.Hist}
\alias{head.Hist}
\title{Check the head of an event history object}
\usage{
\method{head}{Hist}(x)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}
}
\description{
The object is evaluated as a matrix
}
