% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{is.na.Hist}
\alias{is.na.Hist}
\title{Check for missing values in event history object}
\usage{
\method{is.na}{Hist}(x)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}
}
\description{
The object has usually two or three columns and
this function checks each column separately
}
