% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveOneOut.R
\name{leaveOneOut}
\alias{leaveOneOut}
\alias{leaveOneOut.survival}
\alias{leaveOneOut.competing.risks}
\title{Compute jackknife pseudo values.}
\usage{
leaveOneOut(object, times, cause, lag = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{"prodlim"}.}

\item{times}{time points at which to compute leave-one-out
event/survival probabilities.}

\item{cause}{Character (other classes are converted with \code{as.character}).
For competing risks the cause of interest.}

\item{lag}{For survival models only. If \code{TRUE} lag the result, i.e. compute
S(t-) instead of S(t).}

\item{...}{not used}
}
\description{
Compute leave-one-out estimates
}
\details{
This function is the work-horse for \code{jackknife}
}
\seealso{
\code{\link{jackknife}}
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}
