% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/List2Matrix.R
\name{List2Matrix}
\alias{List2Matrix}
\title{Reduce list to a matrix or data.frame with names as new columns}
\usage{
List2Matrix(list, depth, names)
}
\arguments{
\item{list}{A named list which contains nested lists}

\item{depth}{The depth in the list hierarchy until an rbindable object}

\item{names}{Names for the list variables}
}
\value{
Matrix or data.frame.
}
\description{
This function is used by summary.prodlim to deal with results.
}
\details{
Reduction is done with rbind.
}
\examples{
x=list(a=data.frame(u=1,b=2,c=3),b=data.frame(u=3,b=4,c=6))
List2Matrix(x,depth=1,"X")
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}

