% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.prodlim.R
\name{print.prodlim}
\alias{print.prodlim}
\alias{print.neighborhood}
\alias{print.Hist}
\title{Print objects in the prodlim library}
\usage{
\method{print}{prodlim}(x, ...)
}
\arguments{
\item{x}{Object of class \code{prodlim}, \code{Hist} and
\code{neighborhood}.}

\item{\dots}{Not used.}
}
\description{
Pretty printing of objects created with the functionality of the `prodlim'
library.
}
\seealso{
\code{\link{summary.prodlim}}, \code{\link{predict.prodlim}}
}
\author{
Thomas Gerds <tag@biostat.ku.dk>
}
\keyword{survival}
