% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodlim-package.R
\docType{package}
\name{prodlim-package}
\alias{prodlim-package}
\title{Kaplan-Meier and Aalen-Johansen method}
\description{
Fast and user friendly implementation
of nonparametric estimators for censored
event history (survival) analysis.
}
\keyword{internal}
