% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redist.R
\name{redist}
\alias{redist}
\title{Calculation of Efron's re-distribution to the right algorithm to obtain the
Kaplan-Meier estimate.}
\usage{
redist(time, status)
}
\arguments{
\item{time}{A numeric vector of event times.}

\item{status}{The event status vector takes the value \code{1} for observed events and
the value \code{0} for right censored times.}
}
\value{
Calculations needed to
}
\description{
Calculation of Efron's re-distribution to the right algorithm to obtain the
Kaplan-Meier estimate.
}
\examples{
redist(time=c(.35,0.4,.51,.51,.7,.73),status=c(0,1,1,0,0,1))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
prodlim
}

