\name{sampleCompriskFrame}
\Rdversion{1.1}
\alias{sampleCompriskFrame}
\docType{data}
\title{
Dummy data set used for illustrating the use of some functions in the package.
}
\description{
Dummy data set used for illustrating the use of some functions in the package.
}
\usage{data(sampleCompriskFrame)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{event}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
  }
}
\keyword{datasets}
