% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{str.Hist}
\alias{str.Hist}
\title{Check the structure of an event history object}
\usage{
\method{str}{Hist}(x)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}
}
\description{
The object is evaluated as a matrix
}
