% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{[.Hist}
\alias{[.Hist}
\title{Subset an event history object}
\usage{
\method{[}{Hist}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}

\item{i}{passed on to next method}

\item{j}{passed on to next method}

\item{drop}{passed on to next method}
}
\description{
This method makes sure that the attributes are
preserved
}
