% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{subset.Hist}
\alias{subset.Hist}
\title{Subset an event history object}
\usage{
\method{subset}{Hist}(x, subset, select, drop, ...)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}

\item{subset}{passed on to next method}

\item{select}{passed on to next method}

\item{drop}{passed on to next method}

\item{...}{passed on to the next method}
}
\description{
This method makes sure that the attributes are
preserved
}
