% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{tail.Hist}
\alias{tail.Hist}
\title{Check the tail of an event history object}
\usage{
\method{tail}{Hist}(x)
}
\arguments{
\item{x}{object obtained with \code{\link{Hist}}}
}
\description{
The object is evaluated as a matrix
}
