\docType{data}
\name{clusterTestData}
\alias{clusterTestData}
\title{Data set with a clustered data structure, for testing the product limit
method.}
\format{A data frame with 42 observations on the following 4 variables.
\describe{ \item{list("midtimeX")}{event time} \item{list("eventX")}{event
status} \item{list("patientid")}{patient number}
\item{list("AnyCrownFracture")}{a binary predictor} }}
\description{
  Data set with a clustered data structure, for testing the
  product limit method.
}
\examples{
data(clusterTestData)
## maybe str(clusterTestData) ; plot(clusterTestData) ...
}
\keyword{datasets}

