% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt.R
\name{confInt}
\alias{confInt}
\title{Add point-wise confidence limits to the graphs of Kaplan-Meier and
Aalen-Johansen estimates.}
\usage{
confInt(ci, citype, col, lty, lwd, density = 55, ...)
}
\arguments{
\item{ci}{A \code{data.table} with columns \code{time}, \code{lower} and \code{upper}.}

\item{citype}{If \code{"shadow"} then confidence limits are drawn as colored
shadows.  Otherwise, dotted lines are used to show the upper and lower
confidence limits.}

\item{col}{the colour of the lines.}

\item{lty}{the line type of the lines.}

\item{lwd}{the line thickness of the lines.}

\item{density}{For \code{citype="shadow"}, the density of the shade. Default
is 55 percent.}

\item{\dots}{Further arguments that are passed to the function
\code{segments} if \code{type=="bars"} and to \code{lines} else.}
}
\value{
Nil
}
\description{
This function is invoked and controlled by \code{plot.prodlim}.
}
\details{
This function should not be called directly. The arguments can be specified
as \code{Confint.arg} in the call to \code{plot.prodlim}.
}
\seealso{
\code{\link{plot.prodlim}}, \code{\link{atRisk}},
\code{\link{markTime}}
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}
\keyword{survival}
