% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprodlim.R
\name{ggprodlim}
\alias{ggprodlim}
\title{ggplot2::ggplot support for Kaplan-Meier and Aalen-Johansen estimators}
\usage{
ggprodlim(x, x_at, pos_atrisk, ...)
}
\arguments{
\item{x}{object obtained with \code{\link{prodlim}}.}

\item{x_at}{Vector containg the positions for ticks on the x-axis}

\item{pos_atrisk}{Vector with x positions for numbers at risk}

\item{...}{passed on to \code{\link{as.data.table.prodlim}}. Can be used to specify 'cause', 'newdata', and 'times'.}
}
\value{
A ggplot2::ggplot object
}
\description{
ggplot2::ggplot support for Kaplan-Meier and Aalen-Johansen estimators
}
\details{
Important functionality like facet_grid is not yet supported
}
\examples{
library(ggplot2)
d <- SimCompRisk(117)
fit1 <- prodlim(Hist(time,event)~1,data = d)
ggprodlim(fit1,x_at = c(0,4,8,12))
g <- ggprodlim(fit1,x_at = c(0,4,8,12))+ggthemes::theme_solarized()
g+theme(plot.margin = margin(t = 0,r = 0,b = 4,l = 0,"line"))
g <- g+ggplot2::scale_fill_manual(values = grDevices::palette.colors(palette = "Okabe-Ito"))
g <- g+ggplot2::scale_color_manual(values = grDevices::palette.colors(palette = "Okabe-Ito"))
d$X1 <- factor(d$X1,levels=c("1","0"),labels=c("1","0"))
fit <- prodlim(Hist(time,event)~X1,data = d)
ggprodlim(fit,cause = 1,x_at = c(0,6,12))
}
\seealso{
\code{\link{plot.prodlim}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
