\name{print.prodlim}
\alias{print.Hist}
\alias{print.neighborhood}
\alias{print.prodlim}
\title{Print objects in the prodlim library}
\usage{
  \method{print}{prodlim} (x, ...)
}
\arguments{
  \item{x}{Object of class \code{prodlim}, \code{Hist} and
  \code{neighborhood}.}

  \item{\dots}{Not used.}
}
\description{
  Pretty printing of objects created with the functionality
  of the `prodlim' library.
}
\author{
  Thomas Gerds <tag@biostat.ku.dk>
}
\seealso{
  \code{\link{summary.prodlim}},
  \code{\link{predict.prodlim}}
}
\keyword{survival}

