\name{print.prodlim}
\alias{print.prodlim}
\alias{print.neighborhood}
\alias{print.Hist}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print objects in the prodlim library}
\description{
  Pretty printing of objects created with the functionality of the
  `prodlim' library. 
}
\usage{
\method{print}{prodlim}(x, ...)
\method{print}{Hist}(x, ...)
\method{print}{neighborhood}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    Object of class \code{prodlim}, \code{Hist} and \code{neighborhood}.
  }

\item{\dots}{
  Not used.
} 
  
}
% \details{

% }
% \value{
% }
% \references{ ~put references to the literature/web site here ~ }
\author{Thomas Gerds <tag@biostat.ku.dk>}
% \note{ ~~further notes~~ 

%  ~Make other sections like Warning with \section{Warning }{....} ~
% }
\seealso{\code{\link{summary.prodlim}}, \code{\link{predict.prodlim}}}

\keyword{survival}% at least one, from doc/KEYWORDS
