\docType{data}
\name{sampleCompriskFrame}
\alias{sampleCompriskFrame}
\title{Dummy data set used for illustrating the use of some functions in the
package.}
\format{A data frame with 10 observations on the following 4 variables.
\describe{ \item{list("id")}{a numeric vector} \item{list("time")}{a numeric
vector} \item{list("event")}{a numeric vector} \item{list("X")}{a numeric
vector} }}
\description{
  Dummy data set used for illustrating the use of some
  functions in the package.
}
\keyword{datasets}

